EY-2075 Rydding i brevkode + fjerne det som er ubrukt
