+++ apps/etterlatte-brev-api/src/main/kotlin/no/nav/etterlatte/brev/behandling/Personer.kt
@@ -6,4 +6,6 @@ data class Innsender(val navn: String, val fnr: String)
 
 data class Soeker(val navn: String, val fnr: String)
 
+data class Gjenlevende(val navn: String, val fnr: String)
+
 data class Avdoed(val navn: String, val doedsdato: LocalDate)
